# shared/openapi.yaml
openapi: 3.1.0
info:
  title: Athena API
  version: 1.0.0
  description: |
    REST API for the athenaeum server. Served at `/api/v1/` on the same
    origin as the PWA client. Authentication is via httpOnly session cookies.

    Permission flags referenced throughout this spec are a bitmask of the
    constants defined in `server/internal/permissions/permissions.go`. The
    `Administrator` flag is a wildcard that grants every other flag. A user's
    effective permissions are the bitwise OR of all their roles' permission
    bitmasks.
servers:
  - url: /api/v1
    description: Same-origin (PWA client)
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: athenaeum_session
      description: |
        httpOnly session cookie issued by `POST /auth/login` or
        `POST /auth/register`. Required for all authenticated endpoints.
  schemas:
...
    RegisterRequest:
      type: object
      required: [username, password, stay_logged_in]
      properties:
        username:
          type: string
        password:
          type: string
          format: password
        invite_id:
          type: string
          nullable: true
          description: |
            Required when the server is not in first-setup mode (i.e. at least
            one user already exists). Omitted for the first registration,
            which creates the owner account.
        stay_logged_in:
          type: boolean
          description: |
            If true, the session cookie has no expiry. If false, the cookie
            expires after the configured `SessionExpiryDays`.
...
    SetupResponse:
      type: object
      required: [needs_setup, user_count]
      properties:
        needs_setup:
          type: boolean
          description: True when no users exist yet (first-setup mode).
        user_count:
          type: integer
...
  responses:
...
    InternalServerError:
      description: Unexpected server error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
security:
  - cookieAuth: []
paths:
...
  /setup:
    get:
      summary: Setup status
      description: |
        Reports whether the server has any registered users. When
        `needs_setup` is true, the client should route to a first-setup flow
        that creates the owner account (no invite required).
      security: []
      responses:
        '200':
          description: Setup status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SetupResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # --- Auth ---
  /auth/register:
    post:
      summary: Register a new user
      description: |
        Creates a new user and immediately issues a session cookie. The first
        registration on a server with no users becomes the owner and does not
        require an invite; subsequent registrations require a valid `invite_id`.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered and session issued.
          headers:
            Set-Cookie:
              description: |
                `athenaeum_session=<token>; Path=/; HttpOnly; SameSite=Lax`.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/login:
    post:
      summary: Log in
      description: Validates credentials and issues a session cookie.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login succeeded.
          headers:
            Set-Cookie:
              description: |
                `athenaeum_session=<token>; Path=/; HttpOnly; SameSite=Lax`.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/logout:
    post:
      summary: Log out
      description: |
        Invalidates the current session server-side and clears the session
        cookie. Requires authentication.
      responses:
        '200':
          description: Logout succeeded.
          headers:
            Set-Cookie:
              description: |
                `athenaeum_session=; Path=/; HttpOnly; Max-Age=-1` (clears the cookie).
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusMessage'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # --- Users ---
  /users/me:
    get:
      summary: Get the current user
      description: |
        Returns the authenticated user's identity, assigned role names, and
        effective permission bitmask.
      responses:
        '200':
          description: Current user info.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /users:
    get:
      summary: List users
      description: |
        Returns all registered users. Requires the `MANAGE_USERS` permission.
      security:
        - cookieAuth: []
      responses:
        '200':
          description: List of users.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /users/{id}/roles:
    patch:
      summary: Assign roles to a user
      description: |
        Replaces the user's role set with the provided list of role IDs.
        Requires the `MANAGE_USERS` permission.
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignRolesRequest'
      responses:
        '200':
          description: Roles updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusMessage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # --- Invites ---
  /invites:
    post:
      summary: Create an invite
      description: |
        Creates a new invite token with the given number of uses and optional
        expiry. Requires the `MANAGE_USERS` permission.
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInviteRequest'
      responses:
        '201':
          description: Invite created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invite'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'

    get:
      summary: List invites
      description: |
        Returns all invites. Requires the `MANAGE_USERS` permission.
      security:
        - cookieAuth: []
      responses:
        '200':
          description: List of invites.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Invite'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /invites/{id}:
    delete:
      summary: Revoke an invite
      description: |
        Revokes an invite by ID. Requires the `MANAGE_USERS` permission.
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Invite revoked.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusMessage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  # --- Roles ---
  /roles:
    get:
      summary: List roles
      description: |
        Returns all roles. Requires authentication (no specific permission).
      security:
        - cookieAuth: []
      responses:
        '200':
          description: List of roles.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Role'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      summary: Create a role
      description: |
        Creates a new role with the given name, color, position, and
        permission bitmask. Requires the `MANAGE_ROLES` permission.
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRoleRequest'
      responses:
        '201':
          description: Role created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /roles/{id}:
    patch:
      summary: Update a role
      description: |
        Partially updates a role. Only the fields present in the body are
        changed. Requires the `MANAGE_ROLES` permission.
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRoleRequest'
      responses:
        '200':
          description: Role updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusMessage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    delete:
      summary: Delete a role
      description: |
        Deletes a role. Preset roles (Owner, Admin, etc.) cannot be deleted.
        Requires the `MANAGE_ROLES` permission.
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Role deleted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusMessage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  # --- Archives ---
  /archives:
    get:
      summary: List archives
      description: |
        Returns all archives. Requires the `VIEW_MOMENTS` permission.
      security:
        - cookieAuth: []
      responses:
        '200':
          description: List of archives.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Archive'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      summary: Create an archive
      description: |
        Creates a new archive. Requires the `MANAGE_ARCHIVES` permission.
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateArchiveRequest'
      responses:
        '201':
          description: Archive created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Archive'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /archives/{id}:
    patch:
      summary: Update an archive
      description: |
        Renames an archive. Requires the `MANAGE_ARCHIVES` permission.
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateArchiveRequest'
      responses:
        '200':
          description: Archive updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Archive'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    delete:
      summary: Delete an archive
      description: |
        Deletes an archive. Requires the `MANAGE_ARCHIVES` permission.
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Archive deleted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusMessage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  # --- Moments ---
  /moments:
    get:
      summary: List moments
      description: |
        Returns a page of moments, optionally filtered by archive and a
        search query. Results are paginated by `(cursor_ts, cursor_id)`.
        Requires the `VIEW_MOMENTS` permission.
      security:
        - cookieAuth: []
      parameters:
        - name: archive
          in: query
          required: false
          schema:
            type: string
          description: Filter moments to a single archive ID.
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 50
          description: Page size (clamped to 1..500).
        - name: cursor_ts
          in: query
          required: false
          schema:
            type: string
            format: date-time
          description: |
            RFC3339Nano timestamp of the last moment on the previous page.
            Must be supplied together with `cursor_id`.
        - name: cursor_id
          in: query
          required: false
          schema:
            type: string
          description: ID of the last moment on the previous page.
        - name: q
          in: query
          required: false
          schema:
            type: string
          description: Full-text search query over moment title and content.
      responses:
        '200':
          description: List of moments (newest first).
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Moment'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      summary: Create a moment
      description: |
        Creates a new moment in the given archive. Requires the
        `CREATE_MOMENT` permission.
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMomentRequest'
      responses:
        '201':
          description: Moment created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Moment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /tags:
    get:
      summary: List tags
      description: |
        Returns all tags. Requires the `VIEW_MOMENTS` permission.
      security:
        - cookieAuth: []
      responses:
        '200':
          description: List of tags.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tag'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      summary: Create a tag
      description: |
        Creates a new tag. Requires the `MANAGE_TAGS` permission.
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTagRequest'
      responses:
        '201':
          description: Tag created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /tags/{id}:
    patch:
      summary: Update a tag
      description: |
        Partially updates a tag. Requires the `MANAGE_TAGS` permission.
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTagRequest'
      responses:
        '200':
          description: Tag updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    delete:
      summary: Delete a tag
      description: |
        Deletes a tag. Requires the `MANAGE_TAGS` permission.
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Tag deleted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusMessage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  # --- Chat ---
...